Move unix specific logic into _unix.go
