+++ pkg/netns/netns_unix.go
@@ -1,5 +1,7 @@
+// +build !windows
+
 /*
-Copyright 2018 The Containerd Authors.
+Copyright The containerd Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
