+++ pkg/os/os.go
@@ -22,11 +22,7 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/containerd/containerd/mount"
-	"github.com/containerd/fifo"
 	"github.com/docker/docker/pkg/symlink"
-	"golang.org/x/net/context"
-	"golang.org/x/sys/unix"
 )
 
 // OS collects system level operations that need to be mocked out
@@ -34,15 +30,11 @@ import (
 type OS interface {
 	MkdirAll(path string, perm os.FileMode) error
 	RemoveAll(path string) error
-	OpenFifo(ctx context.Context, fn string, flag int, perm os.FileMode) (io.ReadWriteCloser, error)
 	Stat(name string) (os.FileInfo, error)
 	ResolveSymbolicLink(name string) (string, error)
 	FollowSymlinkInScope(path, scope string) (string, error)
 	CopyFile(src, dest string, perm os.FileMode) error
 	WriteFile(filename string, data []byte, perm os.FileMode) error
-	Mount(source string, target string, fstype string, flags uintptr, data string) error
-	Unmount(target string) error
-	LookupMount(path string) (mount.Info, error)
 	Hostname() (string, error)
 }
 
@@ -57,11 +49,6 @@ func (RealOS) MkdirAll(path string, perm os.FileMode) error {
 // RemoveAll will call os.RemoveAll to remove the path and its children.
 func (RealOS) RemoveAll(path string) error {
 	return os.RemoveAll(path)
-}
-
-// OpenFifo will call fifo.OpenFifo to open a fifo.
-func (RealOS) OpenFifo(ctx context.Context, fn string, flag int, perm os.FileMode) (io.ReadWriteCloser, error) {
-	return fifo.OpenFifo(ctx, fn, flag, perm)
 }
 
 // Stat will call os.Stat to get the status of the given file.
@@ -107,33 +94,6 @@ func (RealOS) CopyFile(src, dest string, perm os.FileMode) error {
 // WriteFile will call ioutil.WriteFile to write data into a file.
 func (RealOS) WriteFile(filename string, data []byte, perm os.FileMode) error {
 	return ioutil.WriteFile(filename, data, perm)
-}
-
-// Mount will call unix.Mount to mount the file.
-func (RealOS) Mount(source string, target string, fstype string, flags uintptr, data string) error {
-	return unix.Mount(source, target, fstype, flags, data)
-}
-
-// Unmount will call Unmount to unmount the file.
-func (RealOS) Unmount(target string) error {
-	return Unmount(target)
-}
-
-// LookupMount gets mount info of a given path.
-func (RealOS) LookupMount(path string) (mount.Info, error) {
-	return mount.Lookup(path)
-}
-
-// Unmount unmounts the target. It does not return an error in case the target is not mounted.
-// In case the target does not exist, the appropriate error is returned.
-func Unmount(target string) error {
-	err := unix.Unmount(target, unix.MNT_DETACH)
-	if err == unix.EINVAL {
-		// ignore "not mounted" error
-		err = nil
-	}
-
-	return err
 }
 
 // Hostname will call os.Hostname to get the hostname of the host.
