+++ pkg/server/container_stats_list_unix_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 /*
 Copyright 2018 The containerd Authors.
