+++ pkg/server/container_update_resources_unix.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 /*
 Copyright 2017 The Kubernetes Authors.
