+++ pkg/server/service_windows.go
@@ -1,0 +1,35 @@
+// +build windows
+
+/*
+Copyright The containerd Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package server
+
+import (
+	cni "github.com/containerd/go-cni"
+)
+
+// initPlatform handles linux specific initialization for the CRI service.
+// TODO(windows): Initialize CRI plugin for windows
+func (c *criService) initPlatform() error {
+	return nil
+}
+
+// cniLoadOptions returns cni load options for the windows.
+// TODO(windows): Implement CNI options for windows.
+func (c *criService) cniLoadOptions() []cni.CNIOpt {
+	return nil
+}
